refactor(posts-by-tag): add fetchExternalsByTagName into same file
